Add first probe and beacon attack
Attacks Summary:

Samsung Galaxy S8 Plus / Samsung Galaxy Tab - We were able to see wifi networks on this device using probe responses and beacons. When trying to get a denial of service, we saw that the device does not read all probe responses and beacons and has a limit on the number of wifi networks it shows / has.

Ella's Laptop - We were able to see wifi networks only using beacons. We were only able to see a few wifi networks in a denial of service attack.

Iphone 7 IOS 14.4 - We were unable to see our wifi networks in the wifi settings (using beacons or probe responses).